Replace image uploader with stimulus controller
