+++ assets/js/app.js
@@ -20,7 +20,6 @@ LocalTime.start()
 global.constable = global.constable || {}
 global.constable.commentForm = require('./comment-form')
 global.constable.syntaxHighlighting = require('./syntax-highlighting')
-global.constable.textareaImageUploader = require('./textarea-image-uploader')
 
 import { Application } from "stimulus"
 import { definitionsFromContext } from "stimulus/webpack-helpers"
